scripts/requirements.txt
pandas==2.3.1
python-dotenv==1.1.1
SQLAlchemy==2.0.27
streamlit==1.32.0
